Improvement: Use librato complex measurements.
For `measure.` and `measure#`  metrics, the follow data is produced:
min, median, p95, p99, max, sum, count, mean, last. Currently, this
data is being submitted to Librato as 9 seperate metrics. This can
become expensive. Librato supports a complex metric type. The complex
metric counts as 1 metric when it comes to billing but includes:
min, max, sum, count, mean. This just leaves median, p95, and p99 to
be submitted out of band. It is also reasonable to expect that librato
will accept quantiles in the medium term future.