Feature: Measure L10s from Logplex.
Fixes: #109

Introduced a step in the log packet parsing function
that will check if a log packed is from logplex. If
a packet is from logplex and it matches a regexp that
indicates an L10 and number of messages dropped, we
will log the message from logplex verbatim in
addition to the decrypted librato username. Finally,
we increment a counter in the metchan.