Improvement: Use effecient storage for bucket sum.
As an ongoing improvement for bucket memory utilization
this patch optimizes the way a bucket stores the sum
value. This is specifically designed for the metchan
where we use buckets to aggregate internal metrics.
There are certain metrics for which we only care about the count.
This change makes it possible to not have to store all
of the samples for counters and we simply use a single float value.
This change also impacts the customer bucket flow as it exposes
new methods to interact with bucket opearations.