Improvement: Remove suffix for coutners & samples
Closes #115
Closes #116

There was a time when the Librato interface didn't expose
any sort of complex measurment. There were no fields to put
special calculations. Thus, l2met would suffix all metrics
with the name of the computation. There was also a time when
l2met offered only one way to measure data. Now that we have measure,
sample, and count, it is possible for a metric's calculation type
to be implied.

Furthermore, once Librato supports percentiles in their complex
measurement type, we can remove the median,perc95/99 suffix as well.