Bugfix: Workaround for secret decryption
Related #105
Related #114

There is an outstanding issue (#115) with the pkg that
l2met uses to deal with the encryption of Librato credentials
in which certain characters are not properly parsed which results
in the shared key not being able to decrypt data.

This patch trims the known-to-be-bad characters away from
the generated secret.